requirements.txt
langchain
langsmith
python-pptx
openai==0.28
streamlit
langchain_community
requests
Pillow
